refactor(desktop): extract widget-bar primitives from predictions FeedTab [REL-32]
New desktop/src/components/widget-bar/ — the reusable in-widget control-bar
kit, extracted from the Kalshi/predictions FeedTab (pure move+rename, zero
visual change; predictions swaps to imports):

- Bar.tsx: WidgetBar (sticky @container shell + sentinel/IO pinned
  elevation via callback-ref state), BarDivider, BarPill (ex-LensPill)
- Menu.tsx: useDismiss, MenuPanel, MenuHeading, MenuRow (+ optional
  right-aligned count), FilterTrigger (ex-FilterMenu button)
- Segmented.tsx: Segmented<T> (ex-ViewSwitcher, generalized options)
- SearchBox.tsx: SearchBox (two-stage Esc moved inside) + useSlashFocus
- MultiSelectMenu.tsx: ex-CategoryMenu, generalized noun/options/counts
- SelectMenu.tsx: single-select popover, same anatomy (new)
- GearMenu.tsx: gear trigger + MenuPanel shell (new)

FilterMenu, lenses, cards, roving selection, and the fuzzy matcher stay
private to predictions.

Gate: tsc clean, vitest 490/490, verify-kalshi 58/58, verify-search 25/25.

Part of the kill-the-configure-pages pass (PR 0 of 8). Blocks REL-33..39.
